fix-valid : validaciones corregidas
