--- chef_planner/src/main/java/com/salesianostriana/chefplanner/error/ErrorControllerAdvice.java
@@ -1,54 +1,0 @@
-package com.salesianostriana.chefplanner.error;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@RestControllerAdvice
-public class ErrorControllerAdvice extends ResponseEntityExceptionHandler {
-
-    @ExceptionHandler(BadCredentialsException.class)  // ← Específico para credenciales
-    public ResponseEntity<Map<String, String>> handleBadCredentials(BadCredentialsException ex) {
-        Map<String, String> error = new HashMap<>();
-        error.put("success", "false");
-        error.put("error", "bad_credentials");
-        error.put("message", "Usuario o contraseña incorrectos");  // ← Mensaje amigable
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error);
-    }
-
-    @ExceptionHandler(UsernameNotFoundException.class)  // ← Usuario no existe
-    public ResponseEntity<Map<String, String>> handleUserNotFound(UsernameNotFoundException ex) {
-        Map<String, String> error = new HashMap<>();
-        error.put("success", "false");
-        error.put("error", "user_not_found");
-        error.put("message", "Usuario no encontrado");
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error);
-    }
-
-    // Mantén tu handler genérico como fallback
-    @ExceptionHandler(AuthenticationException.class)
-    public ErrorResponse handleAuthenticationException(AuthenticationException ex) {
-        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(
-                HttpStatus.UNAUTHORIZED, "Error de autenticación"
-        );
-        return ErrorResponse.builder(ex, problemDetail)
-                .header("WWW-Authenticate", "Bearer")
-                .build();
-    }
-
-    @ExceptionHandler(AccessDeniedException.class)
-    public ProblemDetail handleAccessDeniedException(AccessDeniedException ex) {
-        return ProblemDetail.forStatusAndDetail(HttpStatus.FORBIDDEN, ex.getMessage());
-    }
-}
