+++ chef_planner/src/main/resources/messages/errors.properties
@@ -1,1 +1,1 @@
-ingredientRequest.name.uniqueingredient=El nombre del producto no puede estar vacío
+ingredientRequest.name.uniqueingredient=El ingrediente ya existe
